+++ packages/typespec-rust/src/lib.ts
@@ -3,7 +3,7 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { createTypeSpecLibrary, JSONSchemaType } from '@typespec/compiler';
+import { createTypeSpecLibrary, JSONSchemaType, paramMessage } from '@typespec/compiler';
 
 /** the public Rust emitter options */
 export interface RustEmitterOptions {
@@ -32,7 +32,32 @@ const EmitterOptionsSchema: JSONSchemaType<RustEmitterOptions> = {
 
 const libDef = {
   name: '@azure-tools/typespec-rust',
-  diagnostics: {},
+  diagnostics: {
+    'InternalError': {
+      severity: 'error',
+      messages: {
+        default: paramMessage`The emitter encountered an internal error during preprocessing. Please open an issue at https://github.com/Azure/typespec-rust/issues and include the complete error message.\n${'stack'}`
+      }
+    },
+    'InvalidArgument': {
+      severity: 'error',
+      messages: {
+        default: 'Invalid arguments were passed to the emitter.'
+      }
+    },
+    'NameCollision': {
+      severity: 'error',
+      messages: {
+        default: 'The emitter automatically renamed one or more items which resulted in a name collision. Please update the client.tsp to rename the type(s) to avoid the collision.'
+      }
+    },
+    'UnsupportedTsp': {
+      severity: 'error',
+      messages: {
+        default: paramMessage`The emitter encountered a TypeSpec definition that is currently not supported.\n${'stack'}`
+      }
+    }
+  },
   emitter: {
     options: <JSONSchemaType<RustEmitterOptions>>EmitterOptionsSchema,
   },